2009-02-03 18:51 UTC+0100 Przemyslaw Czerpak (druzus/at/priv.onet.pl)   * harbour/source/vm/thread.c     * casting to pacify warning. The code is correct.

+++ harbour/source/vm/thread.c
@@ -917,7 +917,7 @@ static int hb_threadWait( PHB_THREADSTATE * pThreads, int iThreads,
          if( timer <= curr )
             fExit = TRUE;
          else
-            ulMilliSec = timer - curr;
+            ulMilliSec = ( ULONG ) ( timer - curr );
       }
 #endif
